format code, delay test score
