+++ search/application/score/field_decay.py
@@ -1,0 +1,25 @@
+# Copyright (C) 2019-2020 Zilliz. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance
+# with the License. You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed under the License
+# is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+# or implied. See the License for the specific language governing permissions and limitations under the License.
+
+
+from enum import Enum
+
+
+class FieldDecayFunction(Enum):
+    LINEAR = 'linear'
+
+
+def decay_helper(decay_function_mode: FieldDecayFunction):
+    def linear(number, max, min):
+        return list(range(max, min, (max - min) / number))
+
+    helper = {FieldDecayFunction.LINEAR: linear}
+    return helper[decay_function_mode]
